Added file hash verification
Ultimately there was a small logic bug causing files to need to be
reuploaded, so the file hash verification resulted in no failures once
that was fixed. The final write was always passing false for start, but
it was possible to get to that section of code without having performed
a write inside of the read loop.